Fixup tactic to work on all cases
The key idea is to introduce an `EggState`, which keeps:
1. `ix: Nat`, which is a supply of unique IDs.
2. `name2expr: List (Int * Expr)`, which maps these identifiers to
   expressions.

We use these to index rewrites that are available in the context,
and the ones that we create. When we communicate with egg via
the FFI, we stringify these keys. Similarly, when egg returns
data back to us, we parse the strings into these int keys.

The REASON for doing this is to avoid the entire tirefire that
in the previous implementation, we were not updating the `LocalCtx`,
which meant that the new names that we would generate from `MetaM` were not
synced with the `LocalCtx` which is responsible for managing user-facing
names.

Thus, we eschew the use of `LocalCtx`, and we maintain state
ourselves.