WIP in fixing rewrite direction issues
Added a direction when passing rewrites to egg, which we consider when applying them. This is important for the case when the LHS of a RW with mvars has less variables than the RHS. Instead of building a proof of the symmetric equation, we just keep track of the fact that we've sent it reversed to egg and use the same mechanism to reconstruct it from the original rw as we do with the instantiated arguments. The inv-inv example from the group's knuth-bendix theorems still has one issues. It can be seen by changing the direction of invRight/mulOne in this commit: apparently the direction of the rewrites sent to egg does matter, even when saturated...

Added two files, group-inv-inv-{working,broken}-rewrites.json, that only differ in these directions and make the rewrites work or not work.