[WIP] Add example to test scaling.
We have a binary counter that counts upward from 000.
We can read the state at (2^n-1) steps, and check that it does
indeed emit the expected output at the final number of steps.

I imagine this does not work for `egg` since we do not have a single
expression that is repeatedly rewritten. Rather, we have a collection of
"steps" that need to be chained together?