get code to work with untyped variant
I understand what's happening. The code that counts
occurences gets confused by the occurences of variables in the type
domain. This makes the rewrite counts go awry, because what we pass
to rewrite as the 10th  occurrence  is the 10th occurrence in the
TYPED ast, not the UNTYPED AST.

The solution is simple: Prefix everything in the type level with
`type-` so that we don't have a clash.